+++ base.js
@@ -35,7 +35,7 @@ module.exports = {
 		'import/no-named-as-default': 2,
 		'import/no-unresolved': [2, { commonjs: true }],
 		'import/prefer-default-export': 0,
-		indent: [2, 'tab', { SwitchCase: 1 }],
+		'indent': [2, 'tab', { SwitchCase: 1 }],
 		'key-spacing': [2, { mode: 'minimum' }],
 		'keyword-spacing': 2,
 		'lodash/chaining': 0,
@@ -68,10 +68,10 @@ module.exports = {
 		'promise/always-return': 2,
 		'promise/catch-or-return': 2,
 		'promise/param-names': 2,
-		'quote-props': ['warn', 'as-needed'], // Set as major change warning
-		quotes: [2, 'single', { allowTemplateLiterals: true }],
+		'quote-props': ['warn', 'consistent-as-needed'], // Set as major change warning
+		'quotes': [2, 'single', { allowTemplateLiterals: true }],
 		'require-yield': 'warn', // Set as major change warning due to eslint:recommend
-		semi: 2,
+		'semi': 2,
 		'space-before-blocks': [2, 'always'],
 		'space-before-function-paren': [2, 'never'],
 		'space-in-parens': [2, 'never'],